requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0

# Deep Learning
tensorflow>=2.15
keras>=3.0

# Image Processing
opencv-python-headless>=4.8
albumentations>=1.3
Pillow>=10.0

# Medical Imaging
pydicom>=2.4
SimpleITK>=2.3

# Dashboard
streamlit>=1.30
plotly>=5.18

# Export
tf2onnx>=1.16
onnxruntime>=1.16

# Data
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0

# Utilities
tqdm>=4.66
requests>=2.31
